Split server.ts into composition root, app, and entrypoint
server.ts was a god-module: it instantiated every adapter and service,
defined the SyncManager class, wired Express, and started the listener. Worse,
controllers imported `syncManager` from ../server while server imported the
routes that import those controllers — a server → routes → controller → server
import cycle that only worked by load-order luck.

Split into a clean dependency DAG:
- syncManager.ts — domain composition root: adapters, service singletons, the
  SyncManager class, and the `syncManager` instance. No Express, no HTTP.
- app.ts — Express wiring only (basicAuth → json → /api routes), no listener.
- server.ts — entrypoint: startServer (vite/static + listen), process handlers,
  closeServer; re-exports { app, syncManager } for backward compat.
- controllers import syncManager from ../syncManager, breaking the cycle.

Behavior and the public import surface (../server still exports app, syncManager,
closeServer) are unchanged; all 113 tests pass and the bundled prod server boots
and serves /api/health.
